drivers: flash: nrf_qspi_nor: add erase complete polling
Introduce an optional polling mechanism that waits for the WIP bit
to clear before proceeding with the next erase. Parts like the
nrf53 emit an erase started event but expect the client to poll the
target's status. A new property has been added to this device tree
node to control the k_sleep period. This is opt-in, by default the
original behavior will be respected.

Signed-off-by: Cory Todd <cory.todd@motorolasolutions.com>